% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_Artists.R
\name{get_artists}
\alias{get_artists}
\title{Search for and get artists}
\usage{
get_artists(artist_name, return_closest_artist = FALSE,
  accessToken = get_spotify_access_token())
}
\arguments{
\item{accessToken}{Spotify Web API token. Defaults to songwriteR::spotifyAccessTokenGet()}

\item{artistName}{a vector of artist names; can be one or more long.}

\item{artistID}{a vector of unique artist IDs as returned by \code{artistSearch}; can be one or more long.}

\item{returnClosest}{logical; should only the closest matching name be returned. Defaults to \code{TRUE}.}
}
\value{
\code{artistSearch} returns a \code{data.frame} with elements corresponding to the original search terms
the matched names, and the associated unique IDs; \code{artistGet} returns a \code{data.frame} with elements x x x;
}
\description{
These functions search for artist(s) by name (\code{artistSearch}) and unique ID (\code{artistGet}).
}
\examples{
\dontrun{
artistSearch('johnny cash')
}
}
\keyword{artists}
