% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotifyAccessTokenGet.R
\name{spotifyAccessTokenGet}
\alias{spotifyAccessTokenGet}
\title{Retrieve Spotify access token}
\usage{
spotifyAccessTokenGet(clientID = Sys.getenv("SPOTIFY_CLIENT_ID"),
  clientSecret = Sys.getenv("SPOTIFY_CLIENT_SECRET"))
}
\arguments{
\item{clientID}{Defaults to System Envioronment variable "SPOTIFY_CLIENT_ID"}

\item{clientSecret}{Defaults to System Envioronment variable "SPOTIFY_CLIENT_SECRET"}
}
\description{
This function creates a Spotify access token.
}
\examples{
\dontrun{
spotifyAccessTokenGet()
}
}
\keyword{auth}
