% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_Artists.R
\name{artistSearch}
\alias{artistSearch}
\title{Search for and get artists}
\usage{
artistSearch(artistName, returnClosest = TRUE,
  accessToken = spotifyAccessTokenGet())
}
\arguments{
\item{artistName}{a vector of artist names; can be one or more long.}

\item{returnClosest}{logical; should only the closest matching name be returned. Defaults 
to \code{TRUE}.}

\item{accessToken}{Spotify Web API token. Defaults to songwriteR::spotifyAccessTokenGet()}

\item{artistID}{a vector of unique artist IDs as returned by \code{artistSearch}; can be 
one or more long.}
}
\value{
\code{artistSearch} returns a \code{data.frame} with elements corresponding to the 
matched names, and the associated unique IDs; \code{artistGet} returns a \code{data.frame} 
with elements x x x;
}
\description{
These functions search for artist(s) by name (\code{artistSearch}) and unique 
ID (\code{artistGet}).
}
\examples{
\dontrun{
artistSearch('johnny cash')
}
}
\keyword{artists}
